+++ src/Constants.php
@@ -2,9 +2,11 @@
 
 namespace Rubix\ML
 {
-    const TWO_PI = 2. * M_PI;
-
     const EPSILON = 1e-8;
     
     const LOG_EPSILON = -18.420680744;
+
+    const TWO_PI = 2. * M_PI;
+
+    const DEFAULT_WORKERS = 4;
 }